+++ db.lua
@@ -1,7 +1,6 @@
 local defaults = {
     profile = {
         dataVersion = 1,
-        currentSeasonId = 12,
         currentSeasonName = nil,
 
         -- General Settings